app/requirements.txt
alembic==1.14.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
env==0.1.0
executing==2.2.0
fastjsonschema==2.21.1
Flask==3.1.0
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.6.1
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
limits==4.0.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
ordered-set==4.1.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==308; platform_system=="Windows"
pyzmq==26.2.1
qrcode==8.0
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.38
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
yarg==0.1.9
zipp==3.21.0
zxcvbn==4.4.28
